refactor(runs): extract observation and artifact logic into core services (#4382)
